requirements.txt
# Core dependencies
requests==2.31.0
urllib3==2.1.0

# Data processing and statistics
numpy==1.24.3
scipy==1.11.4

# Configuration management
pyyaml==6.0.1

# Environment variables
python-dotenv==1.0.0

# Type checking (development)
mypy==1.7.1
types-requests==2.31.0.10

# Testing (development)
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality (development)
black==23.12.1
flake8==6.1.0
pylint==3.0.3

# Performance monitoring
psutil==5.9.6